Update JakeFridaDemo.py
Added some code meant to track if a fall is currently detected. Until actual fall detection is in place, the best I can do is simulate the beginning/end of a fall, which is done by pressing the 'F' key.
I left comments detailing how that block would have to be changed once fall detection is in.